# Conflicts:
#	air_ticket_booking/build/classes/java/main/com/example/air_ticket_booking/controller/route/RouteController.class
#	air_ticket_booking/build/classes/java/main/com/example/air_ticket_booking/repository/route/IRouteRepository.class
#	air_ticket_booking/src/main/java/com/example/air_ticket_booking/model/account/Account.java

+++ air_ticket_booking/src/main/java/com/example/air_ticket_booking/model/seat/TypeSeat.java
@@ -13,7 +13,7 @@ public class TypeSeat {
     public TypeSeat() {
     }
 
-    public TypeSeat(Long idTypeSeat, String nameTypeSeat, Long priceExtra) {
+    public TypeSeat(Long idTypeSeat, String nameTypeSeat, double priceExtra) {
         this.idTypeSeat = idTypeSeat;
         this.nameTypeSeat = nameTypeSeat;
         this.priceExtra = priceExtra;
